% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Country.R
\docType{data}
\name{Country}
\alias{Country}
\title{List of countries}
\format{
A \code{tiblle} with 4 column, which are :

\describe{

\item{numpays}{A \code{numeric} column. An arbitrary rank of countries}
\item{nompays}{A \code{character} column. The names of countries}
\item{nombourse}{A \code{character} column. Names of stock market}
\item{devise}{A \code{character} column. Official currency of countries}

}
}
\usage{
Country
}
\description{
A simple \code{tibble} of African countries and their stock market
}
\keyword{datasets}
