% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\docType{data}
\name{indices}
\alias{indices}
\title{Main indices of African stock markets}
\format{
A \code{tiblle} with 5 column, which are :

\describe{

\item{nompays}{A \code{character} column. The names of countries}
\item{num_jourI}{A \code{Numeric} column. Data collection day}
\item{nom_ind}{A \code{character} column. Name of stock market indices}
\item{Valeur}{A \code{Numeric} column. Indice value}
\item{Variation}{A \code{Numeric} column. percentage change in indice value}

}
}
\usage{
indices
}
\description{
A simple \code{tibble} on evolution of main indices of African stock markets
}
\keyword{datasets}
