% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_CIV.R
\docType{data}
\name{share_CIV}
\alias{share_CIV}
\title{NIGERIA Stock Exchange Bonds}
\format{
A \code{tiblle} with 8 column, which are :

\describe{

\item{nompays}{A \code{character} column. The names of countries}
\item{nom_entreprise}{A \code{character} column. Name of stock market shares}
\item{num_jourA}{A \code{Numeric} column. Data collection day}
\item{ValeurAction}{A \code{Numeric} column. share value}
\item{VariationAction}{A \code{Numeric} column. percentage change in share value}
\item{VolumeAction}{A \code{Numeric} column. Volume traded}

}
}
\usage{
share_CIV
}
\description{
A simple \code{tibble} on evolution of NIGERIA Exchange Bonds
}
\keyword{datasets}
