% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_MAR.R
\docType{data}
\name{share_MAR}
\alias{share_MAR}
\title{MOROCCO Stock Exchange shares}
\format{
A \code{tiblle} with 6 column, which are :

\describe{

\item{nompays}{A \code{character} column. The names of countries}
\item{nom_entreprise}{A \code{character} column. Name of stock market shares}
\item{num_jourA}{A \code{Numeric} column. Data collection day}
\item{ValeurAction}{A \code{Numeric} column. share value}
\item{VariationAction}{A \code{Numeric} column. percentage change in share value}
\item{VolumeAction}{A \code{Numeric} column. Volume traded}

}
}
\usage{
share_MAR
}
\description{
A simple \code{tibble} on evolution of MOROCCO Exchange shares
}
\keyword{datasets}
